requirements.txt
streamlit==1.28.1
google-generativeai==0.3.2
Pillow==10.0.1
PyPDF2==3.0.1
pdf2image==1.16.3
python-multipart==0.0.6
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
pytest==7.4.3
httpx==0.25.2
